requirements.txt
streamlit
pandas
plotly
mysql-connector-python
openpyxl
